included instancing geometry feature and fixed NaN error
Transform, rotation and scale are implemented for Geometry Instancing. It could certainly be better, but for the moment it does the core tasks.

Another thing I added is a check on NaN errors when adding points. This would crash the script when adding SOP data which for instance has been passed through the copy SOP.